feat(preferences): introduce PidDefinitionViewModel and fix PID list updates
- Refactor PID dialog logic into PidDefinitionViewModel for StateFlow state management.
- Fix UI state synchronization issue where edit and alert threshold updates were not immediately reflected in the RecyclerView.
- Ensure state emissions occur on the main dispatcher and break object equality traps on state updates.
- Ensure user-created PIDs are prioritized at the top of the PID list.
- Fix custom PID deletion logic to filter by explicit ID and properly refresh the list state.
